Broken version screwd up by refactoring
