+++ frontend/sfem_Function.cpp
@@ -84,8 +84,6 @@
 
 #include <map>
 
-#define DUMP_THROUGHPUT 1
-
 namespace sfem {
 
     class Sideset::Impl final {
@@ -1589,2147 +1587,7 @@ namespace sfem {
         ret->impl_->handle_constraints = impl_->handle_constraints;
 
         return ret;
-    }
-
-    class LinearElasticity final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        real_t mu{1}, lambda{1};
-
-        long   calls{0};
-        double total_time{0};
-
-        class Jacobians {
-        public:
-            std::shared_ptr<Buffer<jacobian_t>> adjugate;
-            std::shared_ptr<Buffer<jacobian_t>> determinant;
-
-            Jacobians(const ptrdiff_t n_elements, const int size_adjugate)
-                : adjugate(sfem::create_host_buffer<jacobian_t>(n_elements * size_adjugate)),
-                  determinant(sfem::create_host_buffer<jacobian_t>(n_elements)) {}
-        };
-
-        std::shared_ptr<Jacobians> jacobians;
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("LinearElasticity::create");
-
-            auto mesh = space->mesh_ptr();
-            assert(mesh->spatial_dimension() == space->block_size());
-
-            auto ret = std::make_unique<LinearElasticity>(space);
-
-            real_t SFEM_SHEAR_MODULUS        = 1;
-            real_t SFEM_FIRST_LAME_PARAMETER = 1;
-
-            SFEM_READ_ENV(SFEM_SHEAR_MODULUS, atof);
-            SFEM_READ_ENV(SFEM_FIRST_LAME_PARAMETER, atof);
-
-            ret->mu           = SFEM_SHEAR_MODULUS;
-            ret->lambda       = SFEM_FIRST_LAME_PARAMETER;
-            ret->element_type = (enum ElemType)space->element_type();
-
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::lor_op");
-
-            auto ret          = std::make_shared<LinearElasticity>(space);
-            ret->element_type = macro_type_variant(element_type);
-            ret->mu           = mu;
-            ret->lambda       = lambda;
-            return ret;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::derefine_op");
-
-            auto ret          = std::make_shared<LinearElasticity>(space);
-            ret->element_type = macro_base_elem(element_type);
-            ret->mu           = mu;
-            ret->lambda       = lambda;
-            return ret;
-        }
-
-        const char *name() const override { return "LinearElasticity"; }
-        inline bool is_linear() const override { return true; }
-
-        int initialize() override {
-            auto mesh = space->mesh_ptr();
-
-            if (element_type == HEX8) {
-                jacobians = std::make_shared<Jacobians>(mesh->n_elements(),
-                                                        mesh->spatial_dimension() * elem_manifold_dim(element_type));
-
-                hex8_adjugate_and_det_fill(mesh->n_elements(),
-                                           mesh->elements()->data(),
-                                           mesh->points()->data(),
-                                           jacobians->adjugate->data(),
-                                           jacobians->determinant->data());
-            }
-
-            return SFEM_SUCCESS;
-        }
-
-        LinearElasticity(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        ~LinearElasticity() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("LinearElasticity::apply called %ld times. Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::hessian_crs");
-
-            auto mesh = space->mesh_ptr();
-
-            auto graph = space->node_to_node_graph();
-
-            linear_elasticity_crs_aos(element_type,
-                                      mesh->n_elements(),
-                                      mesh->n_nodes(),
-                                      mesh->elements()->data(),
-                                      mesh->points()->data(),
-                                      this->mu,
-                                      this->lambda,
-                                      graph->rowptr()->data(),
-                                      graph->colidx()->data(),
-                                      values);
-
-            return SFEM_SUCCESS;
-        }
-
-        int hessian_bsr(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::hessian_bsr");
-
-            auto mesh  = space->mesh_ptr();
-            auto graph = space->node_to_node_graph();
-
-            linear_elasticity_bsr(element_type,
-                                  mesh->n_elements(),
-                                  mesh->n_nodes(),
-                                  mesh->elements()->data(),
-                                  mesh->points()->data(),
-                                  this->mu,
-                                  this->lambda,
-                                  graph->rowptr()->data(),
-                                  graph->colidx()->data(),
-                                  values);
-
-            return SFEM_SUCCESS;
-        }
-
-        int hessian_bcrs_sym(const real_t *const  x,
-                             const count_t *const rowptr,
-                             const idx_t *const   colidx,
-                             const ptrdiff_t      block_stride,
-                             real_t **const       diag_values,
-                             real_t **const       off_diag_values) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::hessian_bcrs_sym");
-
-            auto mesh = space->mesh_ptr();
-
-            linear_elasticity_bcrs_sym(element_type,
-                                       mesh->n_elements(),
-                                       mesh->n_nodes(),
-                                       mesh->elements()->data(),
-                                       mesh->points()->data(),
-                                       this->mu,
-                                       this->lambda,
-                                       rowptr,
-                                       colidx,
-                                       block_stride,
-                                       diag_values,
-                                       off_diag_values);
-            return SFEM_SUCCESS;
-        }
-
-        int hessian_block_diag_sym(const real_t *const x, real_t *const values) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::hessian_block_diag_sym");
-
-            auto mesh = space->mesh_ptr();
-            return linear_elasticity_block_diag_sym_aos(element_type,
-                                                        mesh->n_elements(),
-                                                        mesh->n_nodes(),
-                                                        mesh->elements()->data(),
-                                                        mesh->points()->data(),
-                                                        this->mu,
-                                                        this->lambda,
-                                                        values);
-        }
-
-        int hessian_block_diag_sym_soa(const real_t *const x, real_t **const values) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::hessian_block_diag_sym_soa");
-            auto mesh = space->mesh_ptr();
-            return linear_elasticity_block_diag_sym_soa(element_type,
-                                                        mesh->n_elements(),
-                                                        mesh->n_nodes(),
-                                                        mesh->elements()->data(),
-                                                        mesh->points()->data(),
-                                                        this->mu,
-                                                        this->lambda,
-                                                        values);
-        }
-
-        int hessian_diag(const real_t *const, real_t *const out) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::hessian_diag");
-
-            auto mesh = space->mesh_ptr();
-
-            linear_elasticity_assemble_diag_aos(element_type,
-                                                mesh->n_elements(),
-                                                mesh->n_nodes(),
-                                                mesh->elements()->data(),
-                                                mesh->points()->data(),
-                                                this->mu,
-                                                this->lambda,
-                                                out);
-            return SFEM_SUCCESS;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::gradient");
-
-            auto mesh = space->mesh_ptr();
-
-            linear_elasticity_assemble_gradient_aos(element_type,
-                                                    mesh->n_elements(),
-                                                    mesh->n_nodes(),
-                                                    mesh->elements()->data(),
-                                                    mesh->points()->data(),
-                                                    this->mu,
-                                                    this->lambda,
-                                                    x,
-                                                    out);
-
-            return SFEM_SUCCESS;
-        }
-
-        int apply(const real_t *const /*x*/, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::apply");
-
-            auto mesh = space->mesh_ptr();
-
-            double tick = MPI_Wtime();
-
-            if (jacobians) {
-                SFEM_TRACE_SCOPE("linear_elasticity_apply_adjugate_aos");
-                linear_elasticity_apply_adjugate_aos(element_type,
-                                                     mesh->n_elements(),
-                                                     mesh->n_nodes(),
-                                                     mesh->elements()->data(),
-                                                     jacobians->adjugate->data(),
-                                                     jacobians->determinant->data(),
-                                                     this->mu,
-                                                     this->lambda,
-                                                     h,
-                                                     out);
-            } else {
-                SFEM_TRACE_SCOPE("linear_elasticity_apply_aos");
-                linear_elasticity_apply_aos(element_type,
-                                            mesh->n_elements(),
-                                            mesh->n_nodes(),
-                                            mesh->elements()->data(),
-                                            mesh->points()->data(),
-                                            this->mu,
-                                            this->lambda,
-                                            h,
-                                            out);
-            }
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-            calls++;
-
-            return SFEM_SUCCESS;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("LinearElasticity::value");
-
-            auto mesh = space->mesh_ptr();
-
-            linear_elasticity_assemble_value_aos(element_type,
-                                                 mesh->n_elements(),
-                                                 mesh->n_nodes(),
-                                                 mesh->elements()->data(),
-                                                 mesh->points()->data(),
-                                                 this->mu,
-                                                 this->lambda,
-                                                 x,
-                                                 out);
-
-            return SFEM_SUCCESS;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class SemiStructuredLinearElasticity : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        real_t mu{1}, lambda{1};
-        bool   use_affine_approximation{true};  // FIXME the iso-parametric version has probably a bug
-        long   calls{0};
-        double total_time{0};
-
-        ~SemiStructuredLinearElasticity() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("SemiStructuredLinearElasticity[%d]::apply(%s) called %ld times. Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       space->semi_structured_mesh().level(),
-                       use_affine_approximation ? "affine" : "isoparametric",
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("SemiStructuredLinearElasticity::create");
-
-            assert(space->has_semi_structured_mesh());
-            if (!space->has_semi_structured_mesh()) {
-                fprintf(stderr,
-                        "[Error] SemiStructuredLinearElasticity::create requires space with "
-                        "semi_structured_mesh!\n");
-                return nullptr;
-            }
-
-            assert(space->element_type() == SSHEX8);  // REMOVEME once generalized approach
-            auto ret = std::make_unique<SemiStructuredLinearElasticity>(space);
-
-            real_t SFEM_SHEAR_MODULUS        = 1;
-            real_t SFEM_FIRST_LAME_PARAMETER = 1;
-
-            SFEM_READ_ENV(SFEM_SHEAR_MODULUS, atof);
-            SFEM_READ_ENV(SFEM_FIRST_LAME_PARAMETER, atof);
-
-            ret->mu           = SFEM_SHEAR_MODULUS;
-            ret->lambda       = SFEM_FIRST_LAME_PARAMETER;
-            ret->element_type = (enum ElemType)space->element_type();
-
-            int SFEM_HEX8_ASSUME_AFFINE = ret->use_affine_approximation;
-            SFEM_READ_ENV(SFEM_HEX8_ASSUME_AFFINE, atoi);
-            ret->use_affine_approximation = SFEM_HEX8_ASSUME_AFFINE;
-
-            return ret;
-        }
-
-        void set_option(const std::string &name, bool val) override {
-            if (name == "ASSUME_AFFINE") {
-                use_affine_approximation = val;
-            }
-        }
-
-        std::shared_ptr<Op> clone() const override {
-            auto ret = std::make_shared<SemiStructuredLinearElasticity>(space);
-            *ret     = *this;
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            assert(false);
-            fprintf(stderr, "[Error] ss:LinearElasticity::lor_op NOT IMPLEMENTED!\n");
-            return nullptr;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLinearElasticity::derefine_op");
-
-            assert(space->has_semi_structured_mesh() || space->element_type() == macro_base_elem(element_type));
-
-            if (space->has_semi_structured_mesh()) {
-                auto ret                      = std::make_shared<SemiStructuredLinearElasticity>(space);
-                ret->element_type             = element_type;
-                ret->use_affine_approximation = use_affine_approximation;
-                ret->mu                       = mu;
-                ret->lambda                   = lambda;
-                // ret->initialize();
-                return ret;
-            } else {
-                assert(space->element_type() == macro_base_elem(element_type));
-                auto ret          = std::make_shared<LinearElasticity>(space);
-                ret->element_type = macro_base_elem(element_type);
-                ret->mu           = mu;
-                ret->lambda       = lambda;
-                ret->initialize();
-                return ret;
-            }
-        }
-
-        const char *name() const override { return "ss:LinearElasticity"; }
-        inline bool is_linear() const override { return true; }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        SemiStructuredLinearElasticity(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            assert(false);
-            return SFEM_FAILURE;
-        }
-
-        int hessian_bsr(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            auto &ssm = space->semi_structured_mesh();
-            SFEM_TRACE_SCOPE_VARIANT("SemiStructuredLinearElasticity[%d]::hessian_bsr", ssm.level());
-
-            return affine_sshex8_elasticity_bsr(ssm.level(),
-                                                ssm.n_elements(),
-                                                ssm.interior_start(),
-                                                ssm.element_data(),
-                                                ssm.point_data(),
-                                                this->mu,
-                                                this->lambda,
-                                                rowptr,
-                                                colidx,
-                                                values);
-        }
-
-        int hessian_diag(const real_t *const, real_t *const out) override {
-            auto &ssm = space->semi_structured_mesh();
-            SFEM_TRACE_SCOPE_VARIANT("SemiStructuredLinearElasticity[%d]::hessian_diag", ssm.level());
-
-            return affine_sshex8_linear_elasticity_diag(ssm.level(),
-                                                        ssm.n_elements(),
-                                                        ssm.interior_start(),
-                                                        ssm.element_data(),
-                                                        ssm.point_data(),
-                                                        mu,
-                                                        lambda,
-                                                        3,
-                                                        &out[0],
-                                                        &out[1],
-                                                        &out[2]);
-        }
-
-        int hessian_block_diag_sym(const real_t *const x, real_t *const values) override {
-            auto &ssm = space->semi_structured_mesh();
-            SFEM_TRACE_SCOPE_VARIANT("SemiStructuredLinearElasticity[%d]::hessian_block_diag_sym", ssm.level());
-
-            return affine_sshex8_linear_elasticity_block_diag_sym(ssm.level(),
-                                                                  ssm.n_elements(),
-                                                                  ssm.interior_start(),
-                                                                  ssm.element_data(),
-                                                                  ssm.point_data(),
-                                                                  mu,
-                                                                  lambda,
-                                                                  6,
-                                                                  &values[0],
-                                                                  &values[1],
-                                                                  &values[2],
-                                                                  &values[3],
-                                                                  &values[4],
-                                                                  &values[5]);
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override { return apply(nullptr, x, out); }
-
-        int apply(const real_t *const /*x*/, const real_t *const h, real_t *const out) override {
-            auto &ssm = space->semi_structured_mesh();
-            SFEM_TRACE_SCOPE_VARIANT("SemiStructuredLinearElasticity[%d]::apply", ssm.level());
-
-            assert(element_type == SSHEX8);  // REMOVEME once generalized approach
-
-            calls++;
-
-            double tick = MPI_Wtime();
-            int    err;
-            if (use_affine_approximation) {
-                err = affine_sshex8_linear_elasticity_apply(ssm.level(),
-                                                            ssm.n_elements(),
-                                                            ssm.interior_start(),
-                                                            ssm.element_data(),
-                                                            ssm.point_data(),
-                                                            mu,
-                                                            lambda,
-                                                            3,
-                                                            &h[0],
-                                                            &h[1],
-                                                            &h[2],
-                                                            3,
-                                                            &out[0],
-                                                            &out[1],
-                                                            &out[2]);
-
-            } else {
-                err = sshex8_linear_elasticity_apply(ssm.level(),
-                                                     ssm.n_elements(),
-                                                     ssm.interior_start(),
-                                                     ssm.element_data(),
-                                                     ssm.point_data(),
-                                                     mu,
-                                                     lambda,
-                                                     3,
-                                                     &h[0],
-                                                     &h[1],
-                                                     &h[2],
-                                                     3,
-                                                     &out[0],
-                                                     &out[1],
-                                                     &out[2]);
-            }
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-
-            return err;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            assert(false);
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class Laplacian final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        long   calls{0};
-        double total_time{0};
-
-        const char *name() const override { return "Laplacian"; }
-        inline bool is_linear() const override { return true; }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("Laplacian::create");
-
-            assert(1 == space->block_size());
-
-            auto ret          = std::make_unique<Laplacian>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            auto ret          = std::make_shared<Laplacian>(space);
-            ret->element_type = macro_type_variant(element_type);
-            return ret;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            auto ret          = std::make_shared<Laplacian>(space);
-            ret->element_type = macro_base_elem(element_type);
-            return ret;
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        Laplacian(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-        ~Laplacian() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("Laplacian::apply called %ld times. Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_TRACE_SCOPE("Laplacian::hessian_crs");
-
-            auto mesh  = space->mesh_ptr();
-            auto graph = space->dof_to_dof_graph();
-
-            return laplacian_crs(element_type,
-                                 mesh->n_elements(),
-                                 mesh->n_nodes(),
-                                 mesh->elements()->data(),
-                                 mesh->points()->data(),
-                                 graph->rowptr()->data(),
-                                 graph->colidx()->data(),
-                                 values);
-        }
-
-        int hessian_crs_sym(const real_t *const  x,
-                            const count_t *const rowptr,
-                            const idx_t *const   colidx,
-                            real_t *const        diag_values,
-                            real_t *const        off_diag_values) override {
-            SFEM_TRACE_SCOPE("Laplacian::hessian_crs_sym");
-
-            // auto graph = space->node_to_node_graph_upper_triangular();
-
-            auto mesh = space->mesh_ptr();
-
-            return laplacian_crs_sym(element_type,
-                                     mesh->n_elements(),
-                                     mesh->n_nodes(),
-                                     mesh->elements()->data(),
-                                     mesh->points()->data(),
-                                     rowptr,
-                                     colidx,
-                                     diag_values,
-                                     off_diag_values);
-        }
-
-        int hessian_diag(const real_t *const /*x*/, real_t *const values) override {
-            SFEM_TRACE_SCOPE("Laplacian::hessian_diag");
-
-            auto mesh = space->mesh_ptr();
-            return laplacian_diag(
-                    element_type, mesh->n_elements(), mesh->n_nodes(), mesh->elements()->data(), mesh->points()->data(), values);
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("Laplacian::gradient");
-
-            auto mesh = space->mesh_ptr();
-            return laplacian_assemble_gradient(
-                    element_type, mesh->n_elements(), mesh->n_nodes(), mesh->elements()->data(), mesh->points()->data(), x, out);
-        }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("Laplacian::apply");
-
-            auto   mesh = space->mesh_ptr();
-            double tick = MPI_Wtime();
-
-            int err = laplacian_apply(
-                    element_type, mesh->n_elements(), mesh->n_nodes(), mesh->elements()->data(), mesh->points()->data(), h, out);
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-            calls++;
-            return err;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("Laplacian::value");
-
-            auto mesh = space->mesh_ptr();
-            return laplacian_assemble_value(
-                    element_type, mesh->n_elements(), mesh->n_nodes(), mesh->elements()->data(), mesh->points()->data(), x, out);
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class VectorLaplacian final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        std::shared_ptr<Buffer<jacobian_t>> fff;
-
-        long   calls{0};
-        double total_time{0};
-
-        const char *name() const override { return "VectorLaplacian"; }
-        inline bool is_linear() const override { return true; }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("VectorLaplacian::create");
-
-            assert(1 != space->block_size());
-            auto ret          = std::make_unique<VectorLaplacian>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-
-            int SFEM_VECTOR_LAPLACIAN_FFF = 1;
-            SFEM_READ_ENV(SFEM_VECTOR_LAPLACIAN_FFF, atoi);
-
-            if (SFEM_VECTOR_LAPLACIAN_FFF) {
-                ret->fff = create_host_buffer<jacobian_t>(space->mesh_ptr()->n_elements() * 6);
-
-                if (SFEM_SUCCESS != hex8_fff_fill(space->mesh_ptr()->n_elements(),
-                                                  space->mesh_ptr()->elements()->data(),
-                                                  space->mesh_ptr()->points()->data(),
-                                                  ret->fff->data())) {
-                    SFEM_ERROR("Unable to create fff");
-                }
-            }
-
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            auto ret          = std::make_shared<VectorLaplacian>(space);
-            ret->element_type = macro_type_variant(element_type);
-            return ret;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            auto ret          = std::make_shared<VectorLaplacian>(space);
-            ret->element_type = macro_base_elem(element_type);
-            return ret;
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        VectorLaplacian(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-        ~VectorLaplacian() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("VectorLaplacian::apply called %ld times. Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_TRACE_SCOPE("VectorLaplacian::hessian_crs");
-            SFEM_ERROR("IMPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int hessian_crs_sym(const real_t *const  x,
-                            const count_t *const rowptr,
-                            const idx_t *const   colidx,
-                            real_t *const        diag_values,
-                            real_t *const        off_diag_values) override {
-            SFEM_TRACE_SCOPE("VectorLaplacian::hessian_crs_sym");
-            SFEM_ERROR("IMPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int hessian_diag(const real_t *const /*x*/, real_t *const values) override {
-            SFEM_TRACE_SCOPE("VectorLaplacian::hessian_diag");
-            SFEM_ERROR("IMPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("VectorLaplacian::gradient");
-            SFEM_ERROR("IMPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const /*x*/, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("VectorLaplacian::apply");
-
-            double tick = MPI_Wtime();
-
-            const int             block_size = space->block_size();
-            std::vector<real_t *> vec_in(block_size), vec_out(block_size);
-
-            // AoS
-            for (int d = 0; d < block_size; d++) {
-                vec_in[d]  = const_cast<real_t *>(&h[d]);
-                vec_out[d] = &out[d];
-            }
-
-            auto mesh = space->mesh_ptr();
-            int  err;
-            if (this->fff) {
-                err = affine_hex8_vector_laplacian_apply_fff(
-                        // element_type,
-                        mesh->n_elements(),
-                        mesh->elements()->data(),
-                        this->fff->data(),
-                        block_size,
-                        block_size,
-                        vec_in.data(),
-                        vec_out.data());
-            } else {
-                err = affine_hex8_vector_laplacian_apply(
-                        // element_type,
-                        mesh->n_elements(),
-                        mesh->n_nodes(),
-                        mesh->elements()->data(),
-                        mesh->points()->data(),
-                        block_size,
-                        block_size,
-                        vec_in.data(),
-                        vec_out.data());
-            }
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-            calls++;
-            return err;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("VectorLaplacian::value");
-            SFEM_ERROR("IMPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class SemiStructuredVectorLaplacian : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        std::shared_ptr<Buffer<jacobian_t>> fff;
-
-        long   calls{0};
-        double total_time{0};
-
-        ~SemiStructuredVectorLaplacian() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("SemiStructuredVectorLaplacian[%d]::apply called %ld times. "
-                       "Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       space->semi_structured_mesh().level(),
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("SemiStructuredVectorLaplacian::create");
-
-            assert(space->has_semi_structured_mesh());
-            if (!space->has_semi_structured_mesh()) {
-                fprintf(stderr,
-                        "[Error] SemiStructuredVectorLaplacian::create requires space with "
-                        "semi_structured_mesh!\n");
-                return nullptr;
-            }
-
-            assert(space->element_type() == SSHEX8);  // REMOVEME once generalized approach
-            auto ret = std::make_unique<SemiStructuredVectorLaplacian>(space);
-
-            ret->element_type = (enum ElemType)space->element_type();
-
-            ret->fff = create_host_buffer<jacobian_t>(space->mesh_ptr()->n_elements() * 6);
-
-            if (SFEM_SUCCESS != hex8_fff_fill(space->mesh_ptr()->n_elements(),
-                                              space->mesh_ptr()->elements()->data(),
-                                              space->mesh_ptr()->points()->data(),
-                                              ret->fff->data())) {
-                SFEM_ERROR("Unable to create fff");
-            }
-
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            fprintf(stderr, "[Error] ss:Laplacian::lor_op NOT IMPLEMENTED!\n");
-            assert(false);
-            return nullptr;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("SemiStructuredVectorLaplacian::derefine_op");
-
-            assert(space->has_semi_structured_mesh() || space->element_type() == macro_base_elem(element_type));
-            if (space->has_semi_structured_mesh()) {
-                auto ret          = std::make_shared<SemiStructuredVectorLaplacian>(space);
-                ret->element_type = element_type;
-                return ret;
-            } else {
-                auto ret          = std::make_shared<VectorLaplacian>(space);
-                ret->element_type = macro_base_elem(element_type);
-                return ret;
-            }
-        }
-
-        const char *name() const override { return "ss:Laplacian"; }
-        inline bool is_linear() const override { return true; }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        SemiStructuredVectorLaplacian(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_ERROR("[Error] ss:Laplacian::hessian_crs NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int hessian_diag(const real_t *const, real_t *const out) override {
-            SFEM_ERROR("[Error] ss:Laplacian::hessian_diag NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_ERROR("[Error] ss:Laplacian::gradient NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const /*x*/, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SemiStructuredVectorLaplacian::apply");
-
-            assert(element_type == SSHEX8);  // REMOVEME once generalized approach
-
-            auto &ssm = space->semi_structured_mesh();
-
-            double tick = MPI_Wtime();
-
-            const int             block_size = space->block_size();
-            std::vector<real_t *> vec_in(block_size), vec_out(block_size);
-
-            // AoS
-            for (int d = 0; d < block_size; d++) {
-                vec_in[d]  = const_cast<real_t *>(&h[d]);
-                vec_out[d] = &out[d];
-            }
-
-            int err = affine_sshex8_vector_laplacian_apply_fff(ssm.level(),
-                                                               ssm.n_elements(),
-                                                               ssm.element_data(),
-                                                               this->fff->data(),
-                                                               block_size,
-                                                               block_size,
-                                                               vec_in.data(),
-                                                               vec_out.data());
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-            calls++;
-            return err;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_ERROR("[Error] ss:Laplacian::value NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class SemiStructuredLaplacian : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-        bool                           use_affine_approximation{true};
-        bool                           use_stencil{true};
-
-        std::shared_ptr<Buffer<jacobian_t>> fff;
-
-        long   calls{0};
-        double total_time{0};
-
-        ~SemiStructuredLaplacian() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("SemiStructuredLaplacian[%d]::apply(%s) called %ld times. Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       space->semi_structured_mesh().level(),
-                       use_affine_approximation ? (use_stencil ? "stencil" : "affine") : "isoparametric",
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("SemiStructuredLaplacian::create");
-
-            assert(space->has_semi_structured_mesh());
-            if (!space->has_semi_structured_mesh()) {
-                fprintf(stderr,
-                        "[Error] SemiStructuredLaplacian::create requires space with "
-                        "semi_structured_mesh!\n");
-                return nullptr;
-            }
-
-            assert(space->element_type() == SSHEX8);  // REMOVEME once generalized approach
-            auto ret = std::make_unique<SemiStructuredLaplacian>(space);
-
-            ret->element_type = (enum ElemType)space->element_type();
-
-            int SFEM_HEX8_ASSUME_AFFINE = ret->use_affine_approximation;
-            SFEM_READ_ENV(SFEM_HEX8_ASSUME_AFFINE, atoi);
-            ret->use_affine_approximation = SFEM_HEX8_ASSUME_AFFINE;
-
-            int SFEM_ENABLE_HEX8_STENCIL = ret->use_stencil;
-            SFEM_READ_ENV(SFEM_ENABLE_HEX8_STENCIL, atoi);
-            ret->use_stencil = SFEM_ENABLE_HEX8_STENCIL;
-
-            int SFEM_SS_LAPLACIAN_FFF = 1;
-            SFEM_READ_ENV(SFEM_SS_LAPLACIAN_FFF, atoi);
-
-            if (SFEM_SS_LAPLACIAN_FFF) {
-                ret->fff = create_host_buffer<jacobian_t>(space->mesh_ptr()->n_elements() * 6);
-
-                if (SFEM_SUCCESS != hex8_fff_fill(space->mesh_ptr()->n_elements(),
-                                                  space->mesh_ptr()->elements()->data(),
-                                                  space->mesh_ptr()->points()->data(),
-                                                  ret->fff->data())) {
-                    SFEM_ERROR("Unable to create fff");
-                }
-            }
-
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            fprintf(stderr, "[Error] ss:Laplacian::lor_op NOT IMPLEMENTED!\n");
-            assert(false);
-            return nullptr;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLaplacian::derefine_op");
-
-            assert(space->has_semi_structured_mesh() || space->element_type() == macro_base_elem(element_type));
-            if (space->has_semi_structured_mesh()) {
-                auto ret                      = std::make_shared<SemiStructuredLaplacian>(space);
-                ret->element_type             = element_type;
-                ret->use_affine_approximation = use_affine_approximation;
-                ret->use_stencil              = use_stencil;
-                return ret;
-            } else {
-                auto ret          = std::make_shared<Laplacian>(space);
-                ret->element_type = macro_base_elem(element_type);
-                return ret;
-            }
-        }
-
-        const char *name() const override { return "ss:Laplacian"; }
-        inline bool is_linear() const override { return true; }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        SemiStructuredLaplacian(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_ERROR("[Error] ss:Laplacian::hessian_crs NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int hessian_diag(const real_t *const, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLaplacian::hessian_diag");
-
-            auto &ssm = space->semi_structured_mesh();
-            return affine_sshex8_laplacian_diag(
-                    ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), out);
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLaplacian::gradient");
-            return apply(nullptr, x, out);
-        }
-
-        int apply(const real_t *const /*x*/, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLaplacian::apply");
-
-            assert(element_type == SSHEX8);  // REMOVEME once generalized approach
-
-            auto &ssm = space->semi_structured_mesh();
-
-            double tick = MPI_Wtime();
-
-            int err = 0;
-
-            if (this->fff) {
-                SFEM_TRACE_SCOPE("affine_sshex8_laplacian_stencil_apply_fff");
-                affine_sshex8_laplacian_stencil_apply_fff(
-                        ssm.level(), ssm.n_elements(), ssm.element_data(), this->fff->data(), h, out);
-            } else {
-                if (use_stencil) {
-                    err = affine_sshex8_laplacian_stencil_apply(
-                            ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), h, out);
-                } else if (use_affine_approximation) {
-                    err = affine_sshex8_laplacian_apply(
-                            ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), h, out);
-
-                } else {
-                    err = sshex8_laplacian_apply(
-                            ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), h, out);
-                }
-            }
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-            calls++;
-            return err;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_ERROR("[Error] ss:Laplacian::value NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class SpectralElementLaplacian : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        long   calls{0};
-        double total_time{0};
-
-        ~SpectralElementLaplacian() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("SpectralElementLaplacian[%d]::apply called %ld times. Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       space->semi_structured_mesh().level(),
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("SpectralElementLaplacian::create");
-
-            assert(space->has_semi_structured_mesh());
-            if (!space->has_semi_structured_mesh()) {
-                fprintf(stderr,
-                        "[Error] SpectralElementLaplacian::create requires space with "
-                        "semi_structured_mesh!\n");
-                return nullptr;
-            }
-
-            assert(space->element_type() == SSHEX8);  // REMOVEME once generalized approach
-            auto ret          = std::make_unique<SpectralElementLaplacian>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            fprintf(stderr, "[Error] SpectralElementLaplacian::lor_op NOT IMPLEMENTED!\n");
-            assert(false);
-            return nullptr;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("SpectralElementLaplacian::derefine_op");
-
-            assert(space->has_semi_structured_mesh() || space->element_type() == macro_base_elem(element_type));
-            if (space->has_semi_structured_mesh()) {
-                auto ret          = std::make_shared<SpectralElementLaplacian>(space);
-                ret->element_type = element_type;
-                return ret;
-            } else {
-                auto ret          = std::make_shared<Laplacian>(space);
-                ret->element_type = macro_base_elem(element_type);
-                return ret;
-            }
-        }
-
-        const char *name() const override { return "ss:SpectralElementLaplacian"; }
-        inline bool is_linear() const override { return true; }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        SpectralElementLaplacian(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_ERROR("[Error] SpectralElementLaplacian::hessian_crs NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int hessian_diag(const real_t *const, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SpectralElementLaplacian::hessian_diag");
-            SFEM_ERROR("[Error] SpectralElementLaplacian::hessian_diag NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_ERROR("[Error] SpectralElementLaplacian::gradient NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const /*x*/, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SpectralElementLaplacian::apply");
-
-            assert(element_type == SSHEX8);  // REMOVEME once generalized approach
-
-            auto &ssm = space->semi_structured_mesh();
-
-            double tick = MPI_Wtime();
-
-            int err = spectral_hex_laplacian_apply(
-                    ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), h, out);
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-            calls++;
-            return err;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_ERROR("[Error] SpectralElementLaplacian::value NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class SemiStructuredEMLaplacian : public Op {
-    public:
-        std::shared_ptr<FunctionSpace>    space;
-        enum ElemType                     element_type { INVALID };
-        std::shared_ptr<Buffer<scalar_t>> element_matrix;
-
-        long   calls{0};
-        double total_time{0};
-
-        ~SemiStructuredEMLaplacian() {
-            if (DUMP_THROUGHPUT && calls) {
-                printf("SemiStructuredEMLaplacian[%d]::apply() called %ld times. Total: %g [s], "
-                       "Avg: %g [s], TP %g [MDOF/s]\n",
-                       space->semi_structured_mesh().level(),
-                       calls,
-                       total_time,
-                       total_time / calls,
-                       1e-6 * space->n_dofs() / (total_time / calls));
-            }
-        }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("SemiStructuredEMLaplacian::create");
-
-            assert(space->has_semi_structured_mesh());
-            if (!space->has_semi_structured_mesh()) {
-                fprintf(stderr,
-                        "[Error] SemiStructuredEMLaplacian::create requires space with "
-                        "semi_structured_mesh!\n");
-                return nullptr;
-            }
-
-            assert(space->element_type() == SSHEX8);  // REMOVEME once generalized approach
-            auto ret          = std::make_unique<SemiStructuredEMLaplacian>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            fprintf(stderr, "[Error] SemiStructuredEMLaplacian::lor_op NOT IMPLEMENTED!\n");
-            assert(false);
-            return nullptr;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("SemiStructuredEMLaplacian::derefine_op");
-
-            assert(space->has_semi_structured_mesh() || space->element_type() == macro_base_elem(element_type));
-            if (space->has_semi_structured_mesh()) {
-                auto ret          = std::make_shared<SemiStructuredEMLaplacian>(space);
-                ret->element_type = element_type;
-                // FIXME every level stores a variatin of it with different scaling
-                // It woud be usefull to revisit
-                // ret->element_matrix = element_matrix;
-                ret->initialize();
-                return ret;
-            } else {
-                auto ret          = std::make_shared<Laplacian>(space);
-                ret->element_type = macro_base_elem(element_type);
-                return ret;
-            }
-        }
-
-        const char *name() const override { return "ss:em:Laplacian"; }
-        inline bool is_linear() const override { return true; }
-
-        int initialize() override {
-            auto &ssm      = space->semi_structured_mesh();
-            auto  mesh     = space->mesh_ptr();
-            element_matrix = sfem::create_host_buffer<real_t>(mesh->n_elements() * 64);
-            return sshex8_laplacian_element_matrix(ssm.level(),
-                                                   mesh->n_elements(),
-                                                   mesh->n_nodes(),
-                                                   mesh->elements()->data(),
-                                                   mesh->points()->data(),
-                                                   element_matrix->data());
-
-            // element_matrix->print(std::cout);
-            return SFEM_SUCCESS;
-        }
-
-        SemiStructuredEMLaplacian(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_ERROR("[Error] ss:em:Laplacian::hessian_crs NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int hessian_diag(const real_t *const, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLaplacian::hessian_diag");
-
-            auto &ssm = space->semi_structured_mesh();
-            return affine_sshex8_laplacian_diag(
-                    ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), out);
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_ERROR("[Error] ss:em:Laplacian::gradient NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const /*x*/, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("SemiStructuredEMLaplacian::apply");
-
-            assert(element_type == SSHEX8);  // REMOVEME once generalized approach
-
-            auto &ssm = space->semi_structured_mesh();
-
-            double tick = MPI_Wtime();
-
-            int err = sshex8_stencil_element_matrix_apply(
-                    ssm.level(), ssm.n_elements(), ssm.element_data(), element_matrix->data(), h, out);
-
-            double tock = MPI_Wtime();
-            total_time += (tock - tick);
-            calls++;
-            return err;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_ERROR("[Error] ss:em:Laplacian::value NOT IMPLEMENTED!\n");
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class Mass final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        const char *name() const override { return "Mass"; }
-        inline bool is_linear() const override { return true; }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("Mass::create");
-
-            assert(1 == space->block_size());
-
-            auto ret          = std::make_unique<Mass>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-            return ret;
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        Mass(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_TRACE_SCOPE("Mass::hessian_crs");
-
-            auto mesh  = space->mesh_ptr();
-            auto graph = space->dof_to_dof_graph();
-
-            assemble_mass(element_type,
-                          mesh->n_elements(),
-                          mesh->n_nodes(),
-                          mesh->elements()->data(),
-                          mesh->points()->data(),
-                          graph->rowptr()->data(),
-                          graph->colidx()->data(),
-                          values);
-
-            return SFEM_SUCCESS;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("Mass::gradient");
-
-            auto mesh = space->mesh_ptr();
-
-            apply_mass(element_type,
-                       mesh->n_elements(),
-                       mesh->n_nodes(),
-                       mesh->elements()->data(),
-                       mesh->points()->data(),
-                       1,
-                       x,
-                       1,
-                       out);
-
-            return SFEM_SUCCESS;
-        }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("Mass::apply");
-
-            auto mesh = space->mesh_ptr();
-
-            apply_mass(element_type,
-                       mesh->n_elements(),
-                       mesh->n_nodes(),
-                       mesh->elements()->data(),
-                       mesh->points()->data(),
-                       1,
-                       h,
-                       1,
-                       out);
-
-            return SFEM_SUCCESS;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            // auto mesh = space->mesh_ptr();
-
-            // mass_assemble_value((enum ElemType)space->element_type(),
-            //                     mesh->n_elements(),
-            //                     mesh->n_nodes(),
-            //                     mesh->elements()->data(),
-            //                     mesh->points()->data(),
-            //                     x,
-            //                     out);
-
-            // return SFEM_SUCCESS;
-
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class LumpedMass final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        const char *name() const override { return "LumpedMass"; }
-        inline bool is_linear() const override { return true; }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("LumpedMass::create");
-
-            auto ret          = std::make_unique<LumpedMass>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-            return ret;
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        LumpedMass(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_diag(const real_t *const /*x*/, real_t *const values) override {
-            SFEM_TRACE_SCOPE("LumpedMass::hessian_diag");
-
-            auto mesh = space->mesh_ptr();
-
-            if (space->block_size() == 1) {
-                assemble_lumped_mass(element_type,
-                                     mesh->n_elements(),
-                                     mesh->n_nodes(),
-                                     mesh->elements()->data(),
-                                     mesh->points()->data(),
-                                     values);
-            } else {
-                const ptrdiff_t n_nodes = mesh->n_nodes();
-                real_t         *temp    = (real_t *)calloc(n_nodes, sizeof(real_t));
-                assemble_lumped_mass(
-                        element_type, mesh->n_elements(), n_nodes, mesh->elements()->data(), mesh->points()->data(), temp);
-
-                int bs = space->block_size();
-
-#pragma omp parallel for
-                for (ptrdiff_t i = 0; i < n_nodes; i++) {
-                    for (int b = 0; b < bs; b++) {
-                        values[i * bs + b] += temp[i];
-                    }
-                }
-
-                free(temp);
-            }
-
-            return SFEM_SUCCESS;
-        }
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class SemiStructuredLumpedMass final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        const char *name() const override { return "ss:LumpedMass"; }
-        inline bool is_linear() const override { return true; }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("SemiStructuredLumpedMass::create");
-
-            if (!space->has_semi_structured_mesh()) {
-                SFEM_ERROR(
-                        "[Error] SemiStructuredLumpedMass::create requires space with "
-                        "semi_structured_mesh!\n");
-            }
-
-            assert(space->element_type() == SSHEX8);  // REMOVEME once generalized approach
-            auto ret          = std::make_unique<SemiStructuredLumpedMass>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-
-            return ret;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLumpedMass::derefine_op");
-
-            assert(space->has_semi_structured_mesh() || space->element_type() == macro_base_elem(element_type));
-            if (space->has_semi_structured_mesh()) {
-                auto ret          = std::make_shared<SemiStructuredLumpedMass>(space);
-                ret->element_type = element_type;
-                return ret;
-            } else {
-                auto ret          = std::make_shared<LumpedMass>(space);
-                ret->element_type = macro_base_elem(element_type);
-                return ret;
-            }
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        SemiStructuredLumpedMass(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_diag(const real_t *const /*x*/, real_t *const values) override {
-            SFEM_TRACE_SCOPE("SemiStructuredLumpedMass::hessian_diag");
-
-            auto &ssm = space->semi_structured_mesh();
-            if (space->block_size() == 1) {
-                return affine_sshex8_mass_lumped(
-                        ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), values);
-            } else {
-                const ptrdiff_t n = space->n_dofs() / space->block_size();
-
-                auto    buff = create_host_buffer<real_t>(n);
-                real_t *temp = buff->data();
-                int     err  = affine_sshex8_mass_lumped(
-                        ssm.level(), ssm.n_elements(), ssm.interior_start(), ssm.element_data(), ssm.point_data(), temp);
-
-                if (err) SFEM_ERROR("Failure in affine_sshex8_mass_lumped\n");
-
-                int bs = space->block_size();
-#pragma omp parallel for
-                for (ptrdiff_t i = 0; i < n; i++) {
-                    for (int b = 0; b < bs; b++) {
-                        values[i * bs + b] += temp[i];
-                    }
-                }
-            }
-
-            return SFEM_SUCCESS;
-        }
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_ERROR("IMEPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_ERROR("IMEPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            SFEM_ERROR("IMEPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_ERROR("IMEPLEMENT ME!\n");
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class CVFEMMass final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        const char *name() const override { return "CVFEMMass"; }
-        inline bool is_linear() const override { return true; }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            assert(1 == space->block_size());
-
-            auto ret          = std::make_unique<CVFEMMass>(space);
-            ret->element_type = (enum ElemType)space->element_type();
-            return ret;
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        CVFEMMass(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_diag(const real_t *const /*x*/, real_t *const values) override {
-            auto mesh = space->mesh_ptr();
-            cvfem_cv_volumes(
-                    element_type, mesh->n_elements(), mesh->n_nodes(), mesh->elements()->data(), mesh->points()->data(), values);
-
-            return SFEM_SUCCESS;
-        }
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class CVFEMUpwindConvection final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace>  space;
-        std::shared_ptr<Buffer<real_t>> vel[3];
-        enum ElemType                   element_type { INVALID };
-
-        const char *name() const override { return "CVFEMUpwindConvection"; }
-        inline bool is_linear() const override { return true; }
-
-        void set_field(const char * /* name  = velocity */,
-                       const std::shared_ptr<Buffer<real_t>> &v,
-                       const int                              component) override {
-            vel[component] = v;
-        }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            auto mesh = space->mesh_ptr();
-
-            assert(1 == space->block_size());
-
-            auto ret = std::make_unique<CVFEMUpwindConvection>(space);
-
-            const char *SFEM_VELX = nullptr;
-            const char *SFEM_VELY = nullptr;
-            const char *SFEM_VELZ = nullptr;
-
-            SFEM_READ_ENV(SFEM_VELX, );
-            SFEM_READ_ENV(SFEM_VELY, );
-            SFEM_READ_ENV(SFEM_VELZ, );
-
-            if (!SFEM_VELX || !SFEM_VELY || (!SFEM_VELZ && space->mesh_ptr()->spatial_dimension() == 3)) {
-                // fprintf(stderr,
-                //         "No input velocity in env: SFEM_VELX=%s\n,SFEM_VELY=%s\n,SFEM_VELZ=%s\n",
-                //         SFEM_VELX,
-                //         SFEM_VELY,
-                //         SFEM_VELZ);
-                ret->element_type = (enum ElemType)space->element_type();
-                return ret;
-            }
-
-            ptrdiff_t nlocal, nglobal;
-
-            real_t *vel0, *vel1, *vel2;
-            if (array_create_from_file(space->mesh_ptr()->comm()->get(), SFEM_VELX, SFEM_MPI_REAL_T, (void **)&vel0, &nlocal, &nglobal) ||
-                array_create_from_file(space->mesh_ptr()->comm()->get(), SFEM_VELY, SFEM_MPI_REAL_T, (void **)&vel1, &nlocal, &nglobal) ||
-                array_create_from_file(space->mesh_ptr()->comm()->get(), SFEM_VELZ, SFEM_MPI_REAL_T, (void **)&vel2, &nlocal, &nglobal)) {
-                fprintf(stderr, "Unable to read input velocity\n");
-                assert(0);
-                return nullptr;
-            }
-
-            ret->vel[0] = sfem::manage_host_buffer<real_t>(nlocal, vel0);
-            ret->vel[1] = sfem::manage_host_buffer<real_t>(nlocal, vel1);
-            ret->vel[2] = sfem::manage_host_buffer<real_t>(nlocal, vel2);
-
-            return ret;
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        CVFEMUpwindConvection(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        ~CVFEMUpwindConvection() {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            // auto mesh = space->mesh_ptr();
-
-            // auto graph = space->dof_to_dof_graph();
-
-            // cvfem_convection_assemble_hessian(element_type,
-            //                            mesh->nelements,
-            //                            mesh->nnodes,
-            //                            mesh->elements,
-            //                            mesh->points,
-            //                            graph->rowptr()->data(),
-            //                            graph->colidx()->data(),
-            //                            values);
-
-            // return SFEM_SUCCESS;
-
-            SFEM_ERROR("IMPLEMENT ME");
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override { return apply(nullptr, x, out); }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            auto mesh = space->mesh_ptr();
-
-            real_t *vel_[3] = {nullptr, nullptr, nullptr};
-
-            for (int i = 0; i < 3; i++) {
-                if (vel[i]) {
-                    vel_[i] = vel[i]->data();
-                }
-            }
-
-            cvfem_convection_apply(element_type,
-                                   mesh->n_elements(),
-                                   mesh->n_nodes(),
-                                   mesh->elements()->data(),
-                                   mesh->points()->data(),
-                                   vel_,
-                                   h,
-                                   out);
-
-            return SFEM_SUCCESS;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            // auto mesh = space->mesh_ptr();
-
-            // cvfem_convection_assemble_value(element_type,
-            //                          mesh->nelements,
-            //                          mesh->nnodes,
-            //                          mesh->elements,
-            //                          mesh->points,
-            //                          x,
-            //                          out);
-
-            // return SFEM_SUCCESS;
-
-            SFEM_ERROR("IMPLEMENT ME");
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    //
-
-    class NeoHookeanOgden final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace> space;
-        enum ElemType                  element_type { INVALID };
-
-        real_t mu{1}, lambda{1};
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace> &space) {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::create");
-
-            assert(space->mesh_ptr()->spatial_dimension() == space->block_size());
-
-            auto ret = std::make_unique<NeoHookeanOgden>(space);
-
-            real_t SFEM_SHEAR_MODULUS        = 1;
-            real_t SFEM_FIRST_LAME_PARAMETER = 1;
-
-            SFEM_READ_ENV(SFEM_SHEAR_MODULUS, atof);
-            SFEM_READ_ENV(SFEM_FIRST_LAME_PARAMETER, atof);
-
-            ret->mu           = SFEM_SHEAR_MODULUS;
-            ret->lambda       = SFEM_FIRST_LAME_PARAMETER;
-            ret->element_type = (enum ElemType)space->element_type();
-            return ret;
-        }
-
-        std::shared_ptr<Op> lor_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::lor_op");
-
-            auto ret          = std::make_shared<NeoHookeanOgden>(space);
-            ret->element_type = macro_type_variant(element_type);
-            ret->mu           = mu;
-            ret->lambda       = lambda;
-            return ret;
-        }
-
-        std::shared_ptr<Op> derefine_op(const std::shared_ptr<FunctionSpace> &space) override {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::derefine_op");
-
-            auto ret          = std::make_shared<NeoHookeanOgden>(space);
-            ret->element_type = macro_base_elem(element_type);
-            ret->mu           = mu;
-            ret->lambda       = lambda;
-            return ret;
-        }
-
-        const char *name() const override { return "NeoHookeanOgden"; }
-        inline bool is_linear() const override { return true; }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        NeoHookeanOgden(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::hessian_crs");
-
-            auto mesh  = space->mesh_ptr();
-            auto graph = space->node_to_node_graph();
-
-            return neohookean_ogden_hessian_aos(element_type,
-                                                mesh->n_elements(),
-                                                mesh->n_nodes(),
-                                                mesh->elements()->data(),
-                                                mesh->points()->data(),
-                                                this->mu,
-                                                this->lambda,
-                                                x,
-                                                graph->rowptr()->data(),
-                                                graph->colidx()->data(),
-                                                values);
-        }
-
-        int hessian_diag(const real_t *const x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::hessian_diag");
-
-            auto mesh = space->mesh_ptr();
-
-            return neohookean_ogden_diag_aos(element_type,
-                                             mesh->n_elements(),
-                                             mesh->n_nodes(),
-                                             mesh->elements()->data(),
-                                             mesh->points()->data(),
-                                             this->mu,
-                                             this->lambda,
-                                             x,
-                                             out);
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::gradient");
-
-            auto mesh = space->mesh_ptr();
-
-            return neohookean_ogden_gradient_aos(element_type,
-                                                 mesh->n_elements(),
-                                                 mesh->n_nodes(),
-                                                 mesh->elements()->data(),
-                                                 mesh->points()->data(),
-                                                 this->mu,
-                                                 this->lambda,
-                                                 x,
-                                                 out);
-        }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::apply");
-
-            auto mesh = space->mesh_ptr();
-
-            return neohookean_ogden_apply_aos(element_type,
-                                              mesh->n_elements(),
-                                              mesh->n_nodes(),
-                                              mesh->elements()->data(),
-                                              mesh->points()->data(),
-                                              this->mu,
-                                              this->lambda,
-                                              x,
-                                              h,
-                                              out);
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            SFEM_TRACE_SCOPE("NeoHookeanOgden::value");
-
-            auto mesh = space->mesh_ptr();
-
-            return neohookean_ogden_value_aos(element_type,
-                                              mesh->n_elements(),
-                                              mesh->n_nodes(),
-                                              mesh->elements()->data(),
-                                              mesh->points()->data(),
-                                              this->mu,
-                                              this->lambda,
-                                              x,
-                                              out);
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class BoundaryMass final : public Op {
-    public:
-        std::shared_ptr<FunctionSpace>   space;
-        std::shared_ptr<Buffer<idx_t *>> boundary_elements;
-        enum ElemType                    element_type { INVALID };
-
-        const char *name() const override { return "BoundaryMass"; }
-        inline bool is_linear() const override { return true; }
-
-        static std::unique_ptr<Op> create(const std::shared_ptr<FunctionSpace>   &space,
-                                          const std::shared_ptr<Buffer<idx_t *>> &boundary_elements) {
-            SFEM_TRACE_SCOPE("BoundaryMass::create");
-
-            auto ret          = std::make_unique<BoundaryMass>(space);
-            auto element_type = (enum ElemType)space->element_type();
-            ret->element_type = shell_type(side_type(element_type));
-            if (ret->element_type == INVALID) {
-                std::cerr << "Invalid element type for BoundaryMass, Bulk element type: " << type_to_string(element_type) << "\n";
-                return nullptr;
-            }
-            ret->boundary_elements = boundary_elements;
-            return ret;
-        }
-
-        int initialize() override { return SFEM_SUCCESS; }
-
-        BoundaryMass(const std::shared_ptr<FunctionSpace> &space) : space(space) {}
-
-        int hessian_crs(const real_t *const  x,
-                        const count_t *const rowptr,
-                        const idx_t *const   colidx,
-                        real_t *const        values) override {
-            // auto mesh = (mesh_t *)space->mesh().impl_mesh();
-
-            // auto graph = space->dof_to_dof_graph();
-
-            // assemble_mass(element_type,
-            //               boundary_elements->extent(1),
-            //               mesh->nnodes,
-            //               boundary_elements->data(),
-            //               mesh->points,
-            //               graph->rowptr()->data(),
-            //               graph->colidx()->data(),
-            //               values);
-
-            // return SFEM_SUCCESS;
-
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int gradient(const real_t *const x, real_t *const out) override {
-            // auto mesh = (mesh_t *)space->mesh().impl_mesh();
-
-            // assert(1 == space->block_size());
-
-            // apply_mass(element_type,
-            //            boundary_elements->extent(1),
-            //            mesh->nnodes,
-            //            boundary_elements->data(),
-            //            mesh->points,
-            //            x,
-            //            out);
-
-            // return SFEM_SUCCESS;
-
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int apply(const real_t *const x, const real_t *const h, real_t *const out) override {
-            SFEM_TRACE_SCOPE("BoundaryMass::apply");
-
-            auto mesh = space->mesh_ptr();
-
-            int  block_size = space->block_size();
-            auto data       = boundary_elements->data();
-
-            for (int d = 0; d < block_size; d++) {
-                apply_mass(element_type,
-                           boundary_elements->extent(1),
-                           mesh->n_nodes(),
-                           boundary_elements->data(),
-                           mesh->points()->data(),
-                           block_size,
-                           &h[d],
-                           block_size,
-                           &out[d]);
-            }
-
-            return SFEM_SUCCESS;
-        }
-
-        int value(const real_t *x, real_t *const out) override {
-            // auto mesh = (mesh_t *)space->mesh().impl_mesh();
-
-            // mass_assemble_value((enum ElemType)space->element_type(),
-            //                     mesh->nelements,
-            //                     mesh->nnodes,
-            //                     mesh->elements,
-            //                     mesh->points,
-            //                     x,
-            //                     out);
-
-            // return SFEM_SUCCESS;
-
-            assert(0);
-            return SFEM_FAILURE;
-        }
-
-        int report(const real_t *const) override { return SFEM_SUCCESS; }
-    };
-
-    class Factory::Impl {
-    public:
-        std::map<std::string, FactoryFunction>         name_to_create;
-        std::map<std::string, FactoryFunctionBoundary> name_to_create_boundary;
-    };
-
-    Factory::Factory() : impl_(std::make_unique<Impl>()) {}
-
-    Factory::~Factory() = default;
-
-    Factory &Factory::instance() {
-        static Factory instance_;
-
-        if (instance_.impl_->name_to_create.empty()) {
-            instance_.private_register_op("KelvinVoigtNewmark", create_kelvin_voigt_newmark);
-            instance_.private_register_op("LinearElasticity", LinearElasticity::create);
-            instance_.private_register_op("ss:LinearElasticity", SemiStructuredLinearElasticity::create);
-            instance_.private_register_op("Laplacian", Laplacian::create);
-            instance_.private_register_op("VectorLaplacian", VectorLaplacian::create);
-            instance_.private_register_op("ss:VectorLaplacian", SemiStructuredVectorLaplacian::create);
-            instance_.private_register_op("ss:Laplacian", SemiStructuredLaplacian::create);
-            instance_.private_register_op("ss:LumpedMass", SemiStructuredLumpedMass::create);
-            instance_.private_register_op("ss:em:Laplacian", SemiStructuredEMLaplacian::create);
-            instance_.private_register_op("ss:SpectralElementLaplacian", SpectralElementLaplacian::create);
-            instance_.private_register_op("CVFEMUpwindConvection", CVFEMUpwindConvection::create);
-            instance_.private_register_op("Mass", Mass::create);
-            instance_.private_register_op("CVFEMMass", CVFEMMass::create);
-            instance_.private_register_op("LumpedMass", LumpedMass::create);
-            instance_.private_register_op("NeoHookeanOgden", NeoHookeanOgden::create);
-
-            instance_.impl_->name_to_create_boundary["BoundaryMass"] = BoundaryMass::create;
-        }
-
-        return instance_;
-    }
-
-    void Factory::private_register_op(const std::string &name, FactoryFunction factory_function) {
-        impl_->name_to_create[name] = factory_function;
-    }
-
-    void Factory::register_op(const std::string &name, FactoryFunction factory_function) {
-        instance().private_register_op(name, factory_function);
-    }
-
-    std::shared_ptr<Op> Factory::create_op_gpu(const std::shared_ptr<FunctionSpace> &space, const char *name) {
-        return Factory::create_op(space, d_op_str(name).c_str());
-    }
-
-    std::shared_ptr<Op> Factory::create_op(const std::shared_ptr<FunctionSpace> &space, const char *name) {
-        assert(instance().impl_);
-
-        std::string m_name = name;
-
-        if (space->has_semi_structured_mesh()) {
-            m_name = "ss:" + m_name;
-        }
-
-        auto &ntc = instance().impl_->name_to_create;
-        auto  it  = ntc.find(m_name);
-
-        if (it == ntc.end()) {
-            std::cerr << "Unable to find op " << m_name << "\n";
-            return nullptr;
-        }
-
-        return it->second(space);
-    }
-
-    std::shared_ptr<Op> Factory::create_boundary_op(const std::shared_ptr<FunctionSpace>   &space,
-                                                    const std::shared_ptr<Buffer<idx_t *>> &boundary_elements,
-                                                    const char                             *name) {
-        assert(instance().impl_);
-
-        auto &ntc = instance().impl_->name_to_create_boundary;
-        auto  it  = ntc.find(name);
-
-        if (it == ntc.end()) {
-            std::cerr << "Unable to find op " << name << "\n";
-            return nullptr;
-        }
-
-        return it->second(space, boundary_elements);
-    }
-
-    std::string d_op_str(const std::string &name) { return "gpu:" + name; }
+    }  
 
     std::shared_ptr<Buffer<idx_t *>> mesh_connectivity_from_file(const std::shared_ptr<Communicator>& comm, const char *folder) {
         char pattern[1024 * 10];
